Split step error tests and refactor helpers
