+++ crates/rstest-bdd-macros/src/codegen/wrapper/mod.rs
@@ -1,6 +1,7 @@
 //! Wrapper function generation for step definitions.
 
 pub mod args;
+mod arguments;
 pub(crate) mod emit;
 
 #[expect(unused_imports, reason = "re-exports expose helper API")]
